backend/requirements.txt
# FastAPI & ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
supabase==2.0.3
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.0.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# AI & Agents (OpenRouter uses OpenAI-compatible API)
openai>=1.12.0
httpx>=0.24.0

# Background Tasks (optional - can use threading for now)
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing (temporary)
pytest==7.4.3
pytest-asyncio==0.21.1
